# pkg/api/user.go
package api

import (
	"context"
	"errors"
	"fmt"
	"net/http"
	"postui_api/pkg/auth"
	"postui_api/pkg/database"
	"postui_api/pkg/models"

	"gorm.io/gorm"

	"github.com/gin-gonic/gin"
	"golang.org/x/crypto/bcrypt"
)

type UserRepository interface {
	LoginHandler(c *gin.Context)
	RegisterHandler(c *gin.Context)
}

// productRepository holds shared resources like database and Redis client
type userRepository struct {
	DB  database.Database
	Ctx *context.Context
}

func NewUserRepository(db database.Database, ctx *context.Context) *userRepository {
	return &userRepository{
		DB:  db,
		Ctx: ctx,
	}
}

//	@BasePath	/api/v1

// LoginHandler godoc
//	@Summary	Authenticate a user
//	@Schemes
//	@Description	Authenticates a user using username and password, returns a JWT token if successful
//	@Tags			user
//	@Accept			json
//	@Produce		json
//	@Param			user	body		models.LoginUser	true	"User login object"
//	@Success		200		{string}	string				"JWT Token"
//	@Failure		400		{string}	string				"Bad Request"
//	@Failure		401		{string}	string				"Unauthorized"
//	@Failure		500		{string}	string				"Internal Server Error"
//	@Router			/login [post]
func (r *userRepository) LoginHandler(c *gin.Context) {
	var incomingUser models.User
	var dbUser models.User

	// Get JSON body
	if err := c.ShouldBindJSON(&incomingUser); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Bad Request"})
		return
	}

	// Fetch the user from the database
	if err := r.DB.Where("username = ?", incomingUser.Username).First(&dbUser).Error(); err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid username or password"})
		} else {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal Server Error"})
		}
		return
	}

	// Verify password
	if err := bcrypt.CompareHashAndPassword([]byte(dbUser.Password), []byte(incomingUser.Password)); err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid username or password"})
		return
	}

	// Generate JWT token
	token, err := auth.GenerateToken(dbUser.Username)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error generating token"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"token": token})
}

// RegisterHandler godoc
//	@Summary		Register a new user
//	@Schemes		httpdbUser
//	@Description	Registers a new user with the given username and password
//	@Tags			user
//	@Security		JwtAuth
//	@Accept			json
//	@Produce		json
//	@Param			user	body		models.LoginUser	true	"User registration object"
//	@Success		201		{string}	string				"Successfully registered"
//	@Failure		400		{string}	string				"Bad Request"
//	@Failure		500		{string}	string				"Internal Server Error"
//	@Router			/register [post]
func (r *userRepository) RegisterHandler(c *gin.Context) {
	var user models.LoginUser

	if err := c.ShouldBindJSON(&user); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Hash the password
	hashedPassword, err := auth.HashPassword(user.Password)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Could not hash password"})
		return
	}

	// Create new user
	newUser := models.User{Username: user.Username, Password: hashedPassword}

	// Save the user to the database
	if err := r.DB.Create(&newUser).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Could not save user: %v", err)})
		return
	}

	c.JSON(http.StatusCreated, gin.H{"message": "Registration successful"})
}

// ResetPasswordHandler godoc
//	@Summary		Reset user password
//	@Schemes		http
//	@Description	Resets a user password with username and password
//	@Tags			user
//	@Security		JwtAuth
//	@Accept			json
//	@Produce		json
//	@Param			user	body		models.LoginUser	true	"User registration object"
//	@Success		202		{string}	string				"Successfully reset password"
//	@Failure		400		{string}	string				"Bad Request"
//	@Failure		500		{string}	string				"Internal Server Error"
//	@Router			/resetPassword [post]
func (r *userRepository) ResetPasswordHandler(c *gin.Context) {
	var user models.LoginUser
	var dbUser models.User

	if err := c.ShouldBindJSON(&user); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Hash the password
	hashedPassword, err := auth.HashPassword(user.Password)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Could not hash password"})
		return
	}

	// Create new user
	if err := r.DB.Where("username = ?", user.Username).First(&dbUser).Error(); err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			c.JSON(http.StatusUnauthorized, gin.H{"error": "Username don't exists"})
		} else {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal Server Error"})
		}
		return
	}

	dbUser.Password = hashedPassword

	// Save the user to the database
	if err := r.DB.Updates(&dbUser).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Could not save user: %v", err)})
		return
	}

	c.JSON(http.StatusAccepted, gin.H{"message": "Successfully reset password"})
}